refactor(experiences): refactor chart views * add ChartViewDcFilterCount.vue * rename D3Helpers -> d3-helpers, DCHelpers -> dc-helpers
