+++ experiences/components/chart/view/ChartViewLinePieTop.vue
@@ -35,30 +35,9 @@
         </VRow>
       </VCol>
     </ChartViewVRowWebShare>
-    <VRow>
-      <template v-if="filterCount === totalCount">
-        <i18n tag="div" :path="kViewBlock('selected-all')">
-          <template #totalCount>
-            <span class="font-weight-bold" v-text="totalCount" />
-          </template>
-          <template #rowLabel>
-            {{ rowLabel }}
-          </template>
-        </i18n>
-        <span v-t="'click-graph'" />
-      </template>
-      <template v-else>
-        <i18n tag="div" :path="kViewBlock('selected-some')">
-          <template v-for="(v, k) in { filterCount, totalCount }" #[k]>
-            <span :key="k" class="font-weight-bold" v-text="v" />
-          </template>
-          <template #rowLabel>
-            {{ rowLabel }}
-          </template>
-        </i18n>
-        <span>&nbsp;| <a v-t="'Reset All'" @click="resetAll" /></span>
-      </template>
-    </VRow>
+    <ChartViewDcFilterCount
+      v-bind="{ filterCount, totalCount, rowLabel }"
+    />
     <VRow>
       <VCol cols="12">
         <UnitFilterableTable :id="id" v-bind="{ headers: header, items: results }" />
@@ -72,7 +51,7 @@ import * as d3 from 'd3'
 import * as dc from 'dc'
 import crossfilter from 'crossfilter2'
 import mixin from './mixin'
-import { removeEmptyBins } from './utils/DCHelpers'
+import { removeEmptyBins } from './utils/dc-helpers'
 import { datetimeFormatter } from '@/utils/dates'
 
 // Remove warning on default colorscheme, even if not used..
@@ -130,16 +109,10 @@ export default {
         accessor: 'day',
         xUnits: d3.timeDays,
         round: d3.timeDay.round
-      },
-      totalCount: null,
-      filterCount: null
+      }
     }
   },
   methods: {
-    resetAll() {
-      dc.filterAll()
-      dc.renderAll()
-    },
     drawViz() {
       // Define color palette for the graphs
       const colorPalette = [
