+++ experiences/components/chart/view/ChartViewOverviewInstagramMessage.vue
@@ -43,24 +43,9 @@
         </VRow>
       </VCol>
     </ChartViewVRowWebShare>
-    <VRow>
-      <template v-if="filterCount === totalCount">
-        <i18n tag="div" :path="kViewBlock('selected-all')">
-          <template #totalCount>
-            <span class="font-weight-bold" v-text="totalCount" />
-          </template>
-        </i18n>
-        <span v-t="'click-graph'" />
-      </template>
-      <template v-else>
-        <i18n tag="div" :path="kViewBlock('selected-some')">
-          <template v-for="(v, k) in { filterCount, totalCount }" #[k]>
-            <span :key="k" class="font-weight-bold" v-text="v" />
-          </template>
-        </i18n>
-        <span>&nbsp;| <a v-t="'Reset All'" @click="resetAll" /></span>
-      </template>
-    </VRow>
+    <ChartViewDcFilterCount
+      v-bind="{ filterCount, totalCount }"
+    />
     <VRow>
       <VCol cols="12">
         <UnitFilterableTable :id="id" v-bind="{ headers: header, items: results }" />
@@ -74,7 +59,7 @@ import * as d3 from 'd3'
 import * as dc from 'dc'
 import crossfilter from 'crossfilter2'
 import mixin from './mixin'
-import { removeEmptyBins } from './utils/DCHelpers'
+import { removeEmptyBins } from './utils/dc-helpers'
 import { datetimeFormatter } from '@/utils/dates'
 
 // Remove warning on default colorscheme, even if not used..
@@ -93,10 +78,6 @@ export default {
     }
   },
   methods: {
-    resetAll() {
-      dc.filterAll()
-      dc.renderAll()
-    },
     drawViz() {
       // Define color palette for the graphs
       /*
