+++ experiences/components/chart/view/dc/PieChart.vue
@@ -11,7 +11,7 @@
 <script>
 import * as d3 from 'd3'
 import * as dc from 'dc'
-import { addPiePercentage } from '../utils/DCHelpers'
+import { addPiePercentage } from '../utils/dc-helpers'
 import mixin from './mixin'
 
 // Remove warning on default colorscheme, even if not used..
